server-side calls: rename folders within cmk/plugins
