+++ cmk/.f12
@@ -60,7 +60,7 @@ sudo rsync \
 sudo rm -f \
     "$ROOT/lib/python3/cmk/__init__.py" \
     "$ROOT/lib/python3/cmk/active_checks/__init__.py" \
-    "$ROOT/lib/python3/cmk/plugins/collection/config_generation/__init__.py" \
+    "$ROOT/lib/python3/cmk/plugins/collection/server_side_calls/__init__.py" \
     "$ROOT/lib/python3/cmk/special_agents/__init__.py" \
     "$ROOT/lib/python3/cmk/base/__init__.py" \
     "$ROOT/lib/python3/cmk/base/check_legacy_includes/__init__.py" \
