+++ cmk/plugins/collection/server_side_calls/azure.py
