+++ cmk/plugins/collection/server_side_calls/azure_status.py
