+++ tests/unit/cmk/plugins/collection/server_side_calls/test_cisco_meraki.py
@@ -8,7 +8,7 @@ from typing import Any
 
 import pytest
 
-from cmk.plugins.collection.config_generation.cisco_meraki import special_agent_cisco_meraki
+from cmk.plugins.collection.server_side_calls.cisco_meraki import special_agent_cisco_meraki
 from cmk.server_side_calls.v1 import (
     HostConfig,
     HTTPProxy,
