+++ tests/unit/cmk/plugins/mail/server_side_calls/test_mail_loop.py
@@ -7,7 +7,7 @@ from collections.abc import Mapping, Sequence
 
 import pytest
 
-from cmk.plugins.mail.config_generation.mail_loop import active_check_mail_loop
+from cmk.plugins.mail.server_side_calls.mail_loop import active_check_mail_loop
 from cmk.server_side_calls.v1 import HostConfig, IPAddressFamily, PlainTextSecret, StoredSecret
 
 HOST_CONFIG = HostConfig(
